refactor: generic components for restoring and backing up data
