+++ src/components/GoogleAccountsMenu.tsx
@@ -1,7 +1,7 @@
-import { pullGoogleContacts } from '@/actions/integrations'
+import { prepareBackup, pullGoogleContacts, restoreBackup } from '@/actions/integrations'
 import { useNotification } from '@/app/NotificationsProvider'
-import { backupFileData } from '@/lib/definitions'
-import { Popover, Transition } from '@headlessui/react'
+import { BackupFileData } from '@/lib/definitions'
+import { Dialog, Popover, Transition } from '@headlessui/react'
 import {
   CheckCircleIcon,
   ChevronDownIcon,
@@ -10,18 +10,17 @@ import {
 import { CircularProgress } from '@nextui-org/progress'
 import { useEffect, useRef, useState } from 'react'
 import Button from './Button'
-import GoogleAccountBackup from './GoogleAccountBackup'
-import GoogleAccountRestore from './GoogleAccountRestore'
 import LoadingSpinner from './common/spinner'
+import AccountBackup from './BackupDialog'
+import RestoreDialog from './RestoreDialog'
 
 export default function Menu({ googleAccountId }: { googleAccountId: string }) {
-  const { showNotification, hideNotification } = useNotification()
-  const [open, setOpen] = useState(true)
+  const { showNotification } = useNotification()
 
   const [syncing, setSyncing] = useState(false)
 
   const fileInput = useRef<HTMLInputElement>(null)
-  const [backupData, setBackupData] = useState<backupFileData | null>(null)
+  const [backupData, setBackupData] = useState<BackupFileData | null>(null)
 
   const [createdCounter, setCreatedCounter] = useState<number>(0)
   const [createdCounterTotal, setCreatedCounterTotal] = useState<number>(0)
@@ -68,7 +67,7 @@ export default function Menu({ googleAccountId }: { googleAccountId: string }) {
     const content = fileReader.result
     try {
       const parsedData = JSON.parse(content!.toString())
-      let data: backupFileData = {
+      let data: BackupFileData = {
         email: '',
         data: '',
         date: new Date(),
@@ -148,8 +147,8 @@ export default function Menu({ googleAccountId }: { googleAccountId: string }) {
               })
             })
           })
-          ;(await Promise.all(heavyTasks.syncedContactsPromise)) &&
-            (await Promise.all(heavyTasks.createdContactsPromise))
+            ; (await Promise.all(heavyTasks.syncedContactsPromise)) &&
+              (await Promise.all(heavyTasks.createdContactsPromise))
           setSyncing(false)
           setSyncedCounter(0)
           setSyncedCounterTotal(0)
@@ -192,20 +191,55 @@ export default function Menu({ googleAccountId }: { googleAccountId: string }) {
         onChange={handleFileInput}
       ></input>
       {backup && (
-        <GoogleAccountBackup
-          googleAccountId={googleAccountId}
+        <AccountBackup
+          accountId={googleAccountId}
           callClose={async () => showBackup(false)}
-        ></GoogleAccountBackup>
+          dataGetter={prepareBackup}
+          filenamePrefixField='email'
+          title='Backup your google contacts'
+        >
+          <p className="text-sm text-gray-500">
+            Create a backup of your google contacts by clicking the
+            button below. This will download all your contacts in a
+            JSON encrypted file to allow you to restore them later
+            if necessary.
+          </p>
+        </AccountBackup>
       )}
       {restore && (
-        <GoogleAccountRestore
-          googleAccountId={googleAccountId}
+        <RestoreDialog
           callClose={async () => {
             showRestore(false)
             setBackupData(null)
           }}
           backupData={backupData}
-        ></GoogleAccountRestore>
+          restoreFunction={restoreBackup}
+        >
+          {(backupData) => (
+            <>
+              <Dialog.Title
+                as="h3"
+                className="text-base font-semibold leading-6 text-gray-900"
+              >
+                Restore your google contacts backup
+              </Dialog.Title>
+              <div className="mt-2">
+                <p className="text-sm text-gray-500">
+                  Selected backup is from <b>({backupData.email})</b>{' '}
+                  google account made{' '}
+                  <b>{backupData.date.toLocaleString()}</b>. Please
+                  note that this will{' '}
+                  <b>
+                    overwrite all your current contacts in Google
+                    Contacts
+                  </b>
+                  . So be extremely careful while restoring your
+                  contacts.
+                </p>
+              </div>
+            </>
+          )}
+        </RestoreDialog>
       )}
       {fileInput && (
         <Popover className="relative">
